requirement.txt
Flask==2.3.2
torch==2.0.1
Pillow==9.2.0
ultralytics==8.0.196
